Refactor the logic to adjust to the webpage game UI
playGame() is deleted because the game is played
through the webpage directly and not through the
developer tools console.
getHumanChoice() is deleted the player enters their
input by clicking a button from the choices and
not through the prompt.
The rule of the game is changed. The first one to
reach 5 points, wins.
I added selectors and event handlers for the webpage
UI elements.
I refactored playRound() to adjust the logic to the
webpage UI and make the code concise.
I changed announce() to checkWinner() because it has
a better descriptive name.
I added the following functions to help the code easier to understand:
	- checkScore(): verifies if either the player or the computer.
	has reached 5 points.
	- finishGame(): disables the choice buttons.
	- resetGame(): resets the webpage UI to the initial state.
	- startGame(): enables the webpage UI to start the game.
	- loadInitial(): sets the webpage UI to the initial state.