Add alignment styles to the game UI
I added alignment properties to position all the elements in the webpage.
I had to add additional containers to place the elements accurately.
I adjusted the logic for the buttons for reset and start so that the
buttons will display the appropriate button depending on the state of
the game.